requirements.txt
fastapi
uvicorn
opentelemetry-sdk
opentelemetry-exporter-jaeger
